requirements.txt
boto3==1.35.86
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.1
Files.com==1.4.201
idna==3.10
jmespath==1.0.1
numpy==2.2.1
pandas==2.2.3
pyactiveresource==2.2.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
ShopifyAPI==12.7.0
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
